Remove retrying mechanism from code for SQS

+++ src/messaging/Company.Framework.Messaging.Sqs/Consumer/CoreSqsConsumer.cs
@@ -3,9 +3,7 @@ using Amazon.SQS;
 using Amazon.SQS.Model;
 using Company.Framework.Core.Serialization;
 using Company.Framework.Messaging.Consumer;
-using Company.Framework.Messaging.Consumer.Retrying.Args;
 using Company.Framework.Messaging.Sqs.Consumer.Context;
-using Company.Framework.Messaging.Sqs.Consumer.Retrying.Handler;
 using Company.Framework.Messaging.Sqs.Consumer.Settings;
 using Microsoft.Extensions.Logging;
 
@@ -13,11 +11,12 @@ namespace Company.Framework.Messaging.Sqs.Consumer
 {
     public abstract class CoreSqsConsumer<TMessage> : IConsumer
     {
+
+        private static readonly List<string> AttributeNames = new() { "All" };
+
         private readonly SqsConsumerSettings _settings;
-        private readonly SqsConsumerSettings? _retryingSettings;
         private readonly IAmazonSQS _client;
         private readonly ILogger _logger;
-        private readonly ISqsConsumerRetryingHandler? _retryingHandler;
         private readonly IJsonSerializer _jsonSerializer;
 
         private Func<SqsConsumerSettings, CancellationToken, Task> SubscriptionDelegate =>
@@ -29,14 +28,11 @@ namespace Company.Framework.Messaging.Sqs.Consumer
             _client = context.ClientContext.Resolve<IAmazonSQS>();
             _jsonSerializer = context.JsonSerializer;
             _logger = logger;
-            _retryingHandler = context.RetryingHandler;
-            _retryingSettings = _retryingHandler?.ConsumerSettings;
         }
 
         public async Task SubscribeAsync(CancellationToken cancellationToken)
         {
-            await SubscriptionDelegate(_settings, cancellationToken).ConfigureAwait(false);
-            if (_retryingSettings != default) await SubscriptionDelegate(_retryingSettings, cancellationToken).ConfigureAwait(false);
+            await SubscribeToQueue(_settings, cancellationToken).ConfigureAwait(false);
         }
 
         public void Unsubscribe()
@@ -56,7 +52,10 @@ namespace Company.Framework.Messaging.Sqs.Consumer
                 var receiveMessageResponse = await _client.ReceiveMessageAsync(new ReceiveMessageRequest
                 {
                     QueueUrl = queueUrl,
-                    MaxNumberOfMessages = concurrency
+                    AttributeNames = AttributeNames,
+                    MessageAttributeNames = AttributeNames,
+                    MaxNumberOfMessages = concurrency,
+
                 }, cancellationToken).ConfigureAwait(false);
 
                 if (receiveMessageResponse is null || receiveMessageResponse.HttpStatusCode != HttpStatusCode.OK || !receiveMessageResponse.Messages.Any())
@@ -64,46 +63,55 @@ namespace Company.Framework.Messaging.Sqs.Consumer
 
                 foreach (var message in receiveMessageResponse.Messages)
                 {
-                    await _client.DeleteMessageAsync(new DeleteMessageRequest
+                    if (await TryConsumeAsync(message, queueUrl, cancellationToken).ConfigureAwait(false))
                     {
-                        QueueUrl = queueUrl,
-                        ReceiptHandle = message.ReceiptHandle
-                    }, cancellationToken).ConfigureAwait(false);
-
-                    await TryConsumeAsync(message, cancellationToken).ConfigureAwait(false);
+                        await _client.DeleteMessageAsync(new DeleteMessageRequest
+                        {
+                            QueueUrl = queueUrl,
+                            ReceiptHandle = message.ReceiptHandle
+                        }, cancellationToken).ConfigureAwait(false);
+                    }
                 }
             }
         }
 
         private async Task<string> GetQueueUrlAsync(string queue, CancellationToken cancellationToken)
         {
-            var getQueueUrlResponse = await _client.GetQueueUrlAsync(new GetQueueUrlRequest(queue), cancellationToken).ConfigureAwait(false);
-            if (getQueueUrlResponse is null || getQueueUrlResponse.HttpStatusCode != HttpStatusCode.OK)
-                throw new InvalidOperationException($"{queue} is not available for subscription");
-            return getQueueUrlResponse.QueueUrl;
+            try
+            {
+                var getQueueUrlResponse = await _client.GetQueueUrlAsync(new GetQueueUrlRequest(queue), cancellationToken).ConfigureAwait(false);
+                if (getQueueUrlResponse is null || getQueueUrlResponse.HttpStatusCode != HttpStatusCode.OK)
+                    throw new InvalidOperationException($"{queue} is not available for subscription");
+                return getQueueUrlResponse.QueueUrl;
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+                throw;
+            }
+
+
         }
 
-        private async Task TryConsumeAsync(Message message, CancellationToken cancellationToken)
+        private async Task<bool> TryConsumeAsync(Message message, string queueUrl, CancellationToken cancellationToken)
         {
             var typedMessage = _jsonSerializer.Deserialize<TMessage>(message.Body);
-
             if (typedMessage is null)
             {
                 _logger.LogError("Schema is not compatible with message: {}", message.Body);
-                return;
+                return true;
             }
 
             try
             {
                 await ConsumeAsync(typedMessage, cancellationToken);
+                return true;
             }
             catch (Exception exception)
             {
                 _logger.LogError(exception, exception.Message);
-                await Task.Run(() => _retryingHandler?.HandleAsync(new ConsumerRetrialArgs(typedMessage, message.MessageAttributes, exception.GetType()), cancellationToken)
-                        .ConfigureAwait(false), cancellationToken);
+                return false;
             }
         }
-
     }
 }
